feat: Complete kanban tickets MON-002, MON-005, MON-006 (#84)
* feat(tests): update Result<T> API usage patterns and activate tests

- Updated Result<T> API patterns in test files:
  - Changed result.error().code comparison to use static_cast<int>
  - Changed EXPECT_TRUE/FALSE(result) to use .is_ok()/.is_err()

- Activated 7 test files with 135 passing tests:
  - test_result_types.cpp
  - test_di_container.cpp
  - test_monitorable_interface.cpp
  - test_thread_context_simple.cpp
  - test_lock_free_collector.cpp
  - test_distributed_tracing.cpp
  - test_performance_monitoring.cpp

- Updated examples/basic_monitoring_example.cpp for Result<T> API

- Deferred 11 test files requiring missing headers:
  - optimization/, reliability/, utils/ headers not implemented
  - Documented in CMakeLists.txt comments

- Updated MON-002 ticket status to DONE in kanban

* feat(exporters): implement Jaeger/Zipkin HTTP transport

- Added http_transport.h with HTTP client abstraction
  - http_request/http_response structures
  - stub_http_transport for testing
  - simple_http_client for basic HTTP operations

- Updated jaeger_exporter with HTTP transport:
  - Implemented to_thrift_json() for JSON serialization
  - Added send_thrift_batch() with HTTP POST to /api/traces
  - Added retry logic with exponential backoff

- Updated zipkin_exporter with HTTP transport:
  - Implemented to_json_v2() for Zipkin v2 JSON format
  - Added send_json_batch() with HTTP POST to /api/v2/spans
  - Added retry logic with exponential backoff

- Both exporters now support:
  - Custom HTTP headers
  - Configurable timeout
  - Retry with exponential backoff (max 3 retries)

- Updated MON-005 ticket status to DONE in kanban

* docs(kanban): mark MON-006 integration tests as complete

All integration tests are passing (135/135):
- MetricsCollectionTest: Metric registration, counter/gauge operations
- MonitoringIntegrationTest: End-to-end monitoring flow
- MonitoringPerformanceTest: Latency measurements, memory overhead
- ErrorHandlingTest: Error scenarios and recovery

Updated kanban status: 7/7 tickets DONE

+++ tests/test_storage_backends.cpp
@@ -564,7 +564,7 @@ TEST_F(StorageBackendsTest, ErrorHandling) {
         
         auto retrieve_result = backend.retrieve(999);
         EXPECT_FALSE(retrieve_result);
-        EXPECT_EQ(retrieve_result.error().code, monitoring_error_code::not_found);
+        EXPECT_EQ(retrieve_result.error().code, static_cast<int>(monitoring_error_code::not_found));
     }
 }
